apps/worker/requirements.txt
fastapi
uvicorn[standard]
pika
python-dotenv
google-genai
pydantic
